+++ src/main/java/com/example/happy_fitness/controller/CoachController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -27,7 +28,7 @@ public class CoachController {
 	}
 	
 	@GetMapping("{id}")
-	public ResponseEntity<User> getBlogById(@PathVariable("id") Float coachId){
+	public ResponseEntity<User> getBlogById(@PathVariable("id") BigInteger coachId){
 		return new ResponseEntity<User>(coachService.getCoachById(coachId), HttpStatus.OK);
 	}
 }
