+++ src/main/java/com/example/happy_fitness/custom_repository/ProductCustomRepository.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StringUtils;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
@@ -40,7 +41,7 @@ public class ProductCustomRepository {
             "        INNER JOIN" +
             "    facilities f ON f.id = fp.facility_id";
 
-    public List<ProductDto> findProduct(Float facilityId, String status, Float categoryId, Float supplierId, Float minPrice, Float maxPrice) {
+    public List<ProductDto> findProduct(BigInteger facilityId, String status, BigInteger categoryId, BigInteger supplierId, Float minPrice, Float maxPrice) {
         String sql = GET_PRODUCT_QUERY + getCondition(status, categoryId, supplierId, minPrice, maxPrice);
         Query query = entityManager.createNativeQuery(sql, "ProductDto");
         query.setParameter("facilityId", facilityId);
@@ -65,7 +66,7 @@ public class ProductCustomRepository {
         return query.getResultList();
     }
 
-    private String getCondition(String status, Float categoryId, Float supplierId, Float minPrice, Float maxPrice) {
+    private String getCondition(String status, BigInteger categoryId, BigInteger supplierId, Float minPrice, Float maxPrice) {
         String condition = " WHERE f.id = :facilityId";
         if (StringUtils.hasText(status)) {
             condition += " AND fp.status = :status";
@@ -88,7 +89,7 @@ public class ProductCustomRepository {
         return condition;
     }
 
-    public ProductDto findProductDetail(Float facilityId, String code) {
+    public ProductDto findProductDetail(BigInteger facilityId, String code) {
         String sql = GET_PRODUCT_QUERY + " WHERE p.code = :code AND f.id = :facilityId";
         Query query = entityManager.createNativeQuery(sql, "ProductDto");
         query.setParameter("code", code);
