+++ src/main/java/com/example/happy_fitness/repository/BlogRepository.java
@@ -5,9 +5,10 @@ import org.springframework.stereotype.Repository;
 
 import com.example.happy_fitness.entity.Blog;
 
+import java.math.BigInteger;
 
 
 @Repository
-public interface BlogRepository extends JpaRepository<Blog, Long> {
+public interface BlogRepository extends JpaRepository<Blog, BigInteger> {
 
 }
