+++ src/main/java/com/example/happy_fitness/repository/FacilityProductRepository.java
@@ -4,10 +4,11 @@ import com.example.happy_fitness.entity.FacilityProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface FacilityProductRepository extends JpaRepository<FacilityProduct, Float> {
-    List<FacilityProduct> findAllByProduct_Id(Float productId);
-    FacilityProduct findByFacility_IdAndProduct_Id(Float facilityId, Float productId);
+public interface FacilityProductRepository extends JpaRepository<FacilityProduct, BigInteger> {
+    List<FacilityProduct> findAllByProduct_Id(BigInteger productId);
+    FacilityProduct findByFacility_IdAndProduct_Id(BigInteger facilityId, BigInteger productId);
 }
