+++ src/main/java/com/example/happy_fitness/repository/ProductRepository.java
@@ -4,6 +4,8 @@ import com.example.happy_fitness.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Float> {
+public interface ProductRepository extends JpaRepository<Product, BigInteger> {
 }
