+++ src/main/java/com/example/happy_fitness/repository/TicketRepository.java
@@ -4,9 +4,10 @@ import com.example.happy_fitness.entity.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket, Float> {
-    List<Ticket> findAllByFacility_Id(Float facilityId);
+public interface TicketRepository extends JpaRepository<Ticket, BigInteger> {
+    List<Ticket> findAllByFacility_Id(BigInteger facilityId);
 }
