+++ src/main/java/com/example/happy_fitness/service/BaseService.java
@@ -2,12 +2,13 @@ package com.example.happy_fitness.service;
 
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface BaseService<T> {
     T create(UserDetails userDetails, T t);
 
-    String update(T t, Float id, UserDetails userDetails);
+    String update(T t, BigInteger id, UserDetails userDetails);
 
-    void delete(List<Float> ids);
+    void delete(List<BigInteger> ids);
 }
