+++ src/main/java/com/example/happy_fitness/service/BlogService.java
@@ -1,5 +1,6 @@
 package com.example.happy_fitness.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.example.happy_fitness.entity.Blog;
@@ -9,7 +10,7 @@ import com.example.happy_fitness.entity.Blog;
 public interface BlogService {
 	Blog saveBlog(Blog blog);
 	List<Blog> getAllBlog();
-	Blog getBlogById(long id);
-	Blog updateBlog(Blog blog, long id);
-	void deleteBlog(long id);
+	Blog getBlogById(BigInteger id);
+	Blog updateBlog(Blog blog, BigInteger id);
+	void deleteBlog(BigInteger id);
 }
