+++ src/main/java/com/example/happy_fitness/service/impl/CartServiceImpl.java
@@ -14,6 +14,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -33,12 +34,12 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public String update(Cart cart, Float id, UserDetails userDetails) {
+    public String update(Cart cart, BigInteger id, UserDetails userDetails) {
         return null;
     }
 
     @Override
-    public void delete(List<Float> ids) {
+    public void delete(List<BigInteger> ids) {
         cartRepo.deleteAllById(ids);
     }
